+++ src/Clusterers/GaussianMixture.php
@@ -39,7 +39,6 @@ use function exp;
 use function get_object_vars;
 
 use const Rubix\ML\TWO_PI;
-use const Rubix\ML\EPSILON;
 
 /**
  * Gaussian Mixture
@@ -296,7 +295,7 @@ class GaussianMixture implements Estimator, Learner, Probabilistic, Verbose, Per
         ])->check();
 
         if ($this->logger) {
-            $this->logger->info("$this initialized");
+            $this->logger->info("Training $this");
         }
 
         $this->initialize($dataset);
@@ -333,7 +332,7 @@ class GaussianMixture implements Estimator, Learner, Probabilistic, Verbose, Per
 
             if (is_nan($loss)) {
                 if ($this->logger) {
-                    $this->logger->info('Numerical instability detected');
+                    $this->logger->warning('Numerical instability detected');
                 }
 
                 break;
@@ -384,17 +383,25 @@ class GaussianMixture implements Estimator, Learner, Probabilistic, Verbose, Per
 
             $loss /= $n;
 
+            $lossChange = abs($loss - $prevLoss);
+
             $this->losses[$epoch] = $loss;
 
             if ($this->logger) {
-                $this->logger->info("Epoch $epoch - loss: $loss");
+                $lossDirection = $loss < $prevLoss ? '↓' : '↑';
+
+                $message = "Epoch: $epoch, "
+                    . "Loss: $loss, "
+                    . "Loss Change: {$lossDirection}{$lossChange}";
+
+                $this->logger->info($message);
             }
 
             if ($loss <= 0.0) {
                 break;
             }
 
-            if (abs($loss - $prevLoss) < $this->minChange) {
+            if ($lossChange < $this->minChange) {
                 break;
             }
 
@@ -481,7 +488,7 @@ class GaussianMixture implements Estimator, Learner, Probabilistic, Verbose, Per
      * of each of the gaussian components.
      *
      * @param list<int|float> $sample
-     * @return float[]
+     * @return array<int,float>
      */
     protected function jointLogLikelihood(array $sample) : array
     {