+++ tests/Extractors/CSVTest.php
@@ -3,7 +3,7 @@
 namespace Rubix\ML\Tests\Extractors;
 
 use Rubix\ML\Extractors\CSV;
-use Rubix\ML\Extractors\Writable;
+use Rubix\ML\Extractors\Exporter;
 use Rubix\ML\Extractors\Extractor;
 use PHPUnit\Framework\TestCase;
 use IteratorAggregate;
@@ -35,7 +35,7 @@ class CSVTest extends TestCase
     {
         $this->assertInstanceOf(CSV::class, $this->extractor);
         $this->assertInstanceOf(Extractor::class, $this->extractor);
-        $this->assertInstanceOf(Writable::class, $this->extractor);
+        $this->assertInstanceOf(Exporter::class, $this->extractor);
         $this->assertInstanceOf(IteratorAggregate::class, $this->extractor);
         $this->assertInstanceOf(Traversable::class, $this->extractor);
     }
@@ -43,7 +43,19 @@ class CSVTest extends TestCase
     /**
      * @test
      */
-    public function extractWrite() : void
+    public function header() : void
+    {
+        $expected = [
+            'attitude', 'texture', 'sociability', 'rating', 'class',
+        ];
+
+        $this->assertEquals($expected, $this->extractor->header());
+    }
+
+    /**
+     * @test
+     */
+    public function extractExport() : void
     {
         $expected = [
             ['attitude' => 'nice', 'texture' => 'furry', 'sociability' => 'friendly', 'rating' => '4', 'class' => 'not monster'],
@@ -57,6 +69,14 @@ class CSVTest extends TestCase
         $records = iterator_to_array($this->extractor, false);
 
         $this->assertEquals($expected, $records);
+
+        $expected = [
+            'attitude', 'texture', 'sociability', 'rating', 'class',
+        ];
+
+        $header = $this->extractor->header();
+
+        $this->assertEquals($expected, $header);
 
         $this->extractor->export($records);